+++ examples/main/modules.tf
@@ -1,4 +1,4 @@
-module "region" {
+module "azure_region" {
   source  = "claranet/regions/azurerm"
   version = "x.x.x"
 
@@ -9,7 +9,7 @@ module "rg" {
   source  = "claranet/rg/azurerm"
   version = "x.x.x"
 
-  location    = module.region.location
+  location    = module.azure_region.location
   client_name = var.client_name
   environment = var.environment
   stack       = var.stack
@@ -19,10 +19,10 @@ module "logs" {
   source  = "claranet/run/azurerm//modules/logs"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -32,10 +32,10 @@ module "vnet_01" {
   source  = "claranet/vnet/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -49,9 +49,9 @@ module "subnet_01" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -70,10 +70,10 @@ module "vnet_02" {
   source  = "claranet/vnet/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -87,9 +87,9 @@ module "subnet_02" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -110,10 +110,10 @@ module "key_vault" {
   source  = "claranet/keyvault/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -130,10 +130,10 @@ module "lb" {
   source  = "claranet/lb/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -142,8 +142,8 @@ module "lb" {
 }
 
 resource "azurerm_private_link_service" "example" {
-  name     = "pls-${var.stack}-${var.client_name}-${module.region.location_short}-${var.environment}"
-  location = module.region.location
+  name     = format("pls-%s-%s-%s-%s", var.stack, var.client_name, module.azure_region.location_short, var.environment)
+  location = module.azure_region.location
 
   resource_group_name = module.rg.resource_group_name
 
@@ -160,10 +160,10 @@ module "kv_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -189,10 +189,10 @@ module "example_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -210,10 +210,10 @@ module "example_alias_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name