+++ examples/private-dns-zone/modules.tf
@@ -1,4 +1,4 @@
-module "region" {
+module "azure_region" {
   source  = "claranet/regions/azurerm"
   version = "x.x.x"
 
@@ -9,7 +9,7 @@ module "rg" {
   source  = "claranet/rg/azurerm"
   version = "x.x.x"
 
-  location    = module.region.location
+  location    = module.azure_region.location
   client_name = var.client_name
   environment = var.environment
   stack       = var.stack
@@ -19,10 +19,10 @@ module "logs" {
   source  = "claranet/run/azurerm//modules/logs"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -32,10 +32,10 @@ module "vnet" {
   source  = "claranet/vnet/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -47,9 +47,9 @@ module "subnet" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -68,10 +68,10 @@ module "key_vault" {
   source  = "claranet/keyvault/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -101,10 +101,10 @@ module "kv_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name