+++ examples/main/modules.tf
@@ -1,4 +1,4 @@
-module "region" {
+module "azure_region" {
   source  = "claranet/regions/azurerm"
   version = "x.x.x"
 
@@ -9,7 +9,7 @@ module "rg" {
   source  = "claranet/rg/azurerm"
   version = "x.x.x"
 
-  location    = module.region.location
+  location    = module.azure_region.location
   client_name = var.client_name
   environment = var.environment
   stack       = var.stack
@@ -19,10 +19,10 @@ module "logs" {
   source  = "claranet/run/azurerm//modules/logs"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -32,10 +32,10 @@ module "vnet_01" {
   source  = "claranet/vnet/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -49,9 +49,9 @@ module "subnet_01" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -70,10 +70,10 @@ module "vnet_02" {
   source  = "claranet/vnet/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -87,9 +87,9 @@ module "subnet_02" {
   source  = "claranet/subnet/azurerm"
   version = "x.x.x"
 
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -99,7 +99,7 @@ module "subnet_02" {
   virtual_network_name = module.vnet_02.virtual_network_name
 
   private_link_endpoint_enabled = true
-  private_link_service_enabled  = true
+  private_link_service_enabled  = false
 
   subnet_cidr_list = ["172.16.4.0/24"]
 }
@@ -110,10 +110,10 @@ module "key_vault" {
   source  = "claranet/keyvault/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -130,10 +130,10 @@ module "lb" {
   source  = "claranet/lb/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -142,8 +142,8 @@ module "lb" {
 }
 
 resource "azurerm_private_link_service" "example" {
-  name     = "pls-${var.stack}-${var.client_name}-${module.region.location_short}-${var.environment}"
-  location = module.region.location
+  name     = format("pls-%s-%s-%s-%s", var.stack, var.client_name, module.azure_region.location_short, var.environment)
+  location = module.azure_region.location
 
   resource_group_name = module.rg.resource_group_name
 
@@ -160,17 +160,24 @@ module "kv_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
 
   name_suffix = "kv"
 
-  subnet_id        = module.subnet_01.subnet_id
+  custom_private_endpoint_nic_name = "foo"
+
+  subnet_id = module.subnet_01.subnet_id
+  ip_configurations = [{           # The number of IP configurations depends on the target resource
+    member_name        = "default" # The `member_name` value depends on the target resource
+    private_ip_address = cidrhost(module.subnet_01.subnet_cidr_list[0], 34)
+  }]
+
   target_resource  = module.key_vault.key_vault_id
   subresource_name = "vault"
 
@@ -182,17 +189,20 @@ module "example_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
 
   name_suffix = "example"
 
-  subnet_id       = module.subnet_02.subnet_id
+  custom_private_endpoint_nic_name = "bar"
+
+  subnet_id = module.subnet_02.subnet_id
+
   target_resource = azurerm_private_link_service.example.id
 }
 
@@ -200,10 +210,10 @@ module "example_alias_private_endpoint" {
   source  = "claranet/private-endpoint/azurerm"
   version = "x.x.x"
 
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   client_name    = var.client_name
   environment    = var.environment
-  location       = module.region.location
-  location_short = module.region.location_short
   stack          = var.stack
 
   resource_group_name = module.rg.resource_group_name
@@ -212,6 +222,10 @@ module "example_alias_private_endpoint" {
 
   is_manual_connection = true
 
-  subnet_id       = module.subnet_02.subnet_id
+  subnet_id = module.subnet_02.subnet_id
+  ip_configurations = [{
+    private_ip_address = cidrhost(module.subnet_02.subnet_cidr_list[0], 34)
+  }]
+
   target_resource = azurerm_private_link_service.example.alias
 }