+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# Unreleased
+
+Added
+  * AZ-1090: Add the `custom_network_interface_name` parameter to be able to define the custom NIC name of the Private Endpoint
+  * AZ-1090: Add the `ip_configuration` block to be able to define the private IP address (or addresses) of the Private Endpoint
+
+Breaking
+  * AZ-1090: The minimum version of Terraform is now `1.3` to support optional attributes with default values
+
+Changed
+  * AZ-1090: The minimum version of the AzureRM Provider is now `3.36` to support new parameters
+  * AZ-1090: Update example with new parameters
+
+Fixed
+  * AZ-1096: Fix the regex that does not match FHIR IDs
+
 # v6.3.1 - 2023-03-10
 
 Fixed